available.desired.dataFrames(ayrton)	R Documentation

Dataframes for GUI construction

Description

The input of this function are two data.frame, say A and B. The output is A\B and A intersection B.

Usage


available.desired.dataFrames(dataFrame1,dataFrame2,specialCharacter="\r")

Arguments

dataFrame1 	 a data.frame

dataFrame2 	 a data.frame

specialCharacter 	 the character used to split the string used to join all data

Details

This function must be reworked using the merge function.

Value

A list containing two data.frame: dataFrame1=A\B.and dataFrame2=A inters B.

Author(s)

Claudio Ortelli

[Package ayrton version 0.1-0 Index]