get.tclSortModes(ayrton)	R Documentation

Get the sortmodes for tcl types

Description

Given a vector x of tcl types, the functions returns a vector of the same length as x with the corresponding sort modes.

Usage


get.tclSortModes(x)

Arguments

x 	 a vector of tcl types

Details

da fare

Value

A vector of the same length than x containing the corresponding sort modes.

Author(s)

Claudio Ortelli

[Package ayrton version 0.1-0 Index]