getStorageMode(ayrton)	R Documentation

Get the storage mode of a data.frame matrix or vector

Description

This function return the storage modes of the elements of a list (data.frame) or of a matrix or vector.

Usage


getStorageMode(x)

Arguments

x 	 a list, data.frame, matrix or vector

Details

~~ If necessary, more details than the description above ~~

Value

The vector of storage modes. For NULL data frame or list it returns "list".

Author(s)

Claudio Ortelli

[Package ayrton version 0.1-0 Index]